# ✨ Read files content lazily [joelanford edits]