--- a/errors.go
+++ b/errors.go
@@ -11,10 +11,22 @@ var (
 	ErrDir    = errors.New("is a directory")
 )
 
-func newErrNotDir(op, name string) error {
-	return &fs.PathError{Op: op, Path: name, Err: ErrNotDir}
+func newErrNotDir(op, path string) error {
+	return newErr(op, path, ErrNotDir)
 }
 
-func newErrDir(op, name string) error {
-	return &fs.PathError{Op: op, Path: name, Err: ErrDir}
+func newErrDir(op, path string) error {
+	return newErr(op, path, ErrDir)
+}
+
+func newErrClosed(op, path string) error {
+	return newErr(op, path, fs.ErrClosed)
+}
+
+func newErrNotExist(op, path string) error {
+	return newErr(op, path, fs.ErrNotExist)
+}
+
+func newErr(op, path string, err error) error {
+	return &fs.PathError{Op: op, Path: path, Err: err}
 }