--- a/go.mod
+++ b/go.mod
@@ -1,5 +1,11 @@
 module github.com/nlepage/go-tarfs
 
-go 1.16
+go 1.17
 
 require github.com/stretchr/testify v1.7.1
+
+require (
+	github.com/davecgh/go-spew v1.1.0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.0-20200313102051-9f266ea9e77c // indirect
+)