… .github/workflows/go.yml
… README.md
… benchmarks/.gitignore
… example_stat_test.go
… fs_test.go
… go.sum