Based somewhat on https://github.com/nlepage/go-tarfs/issues/15#issuecomment-1646182255, this extra commit is what I was thinking.

Benchmarks before:
```
$ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/nlepage/go-tarfs/benchmarks
BenchmarkOpenTarThenCopyFile_ManySmallFiles-10                        20          57586710 ns/op
BenchmarkOpenTarThenCopyFile_ManySmallFiles_InMemory-10               21          50357206 ns/op
BenchmarkCopyFile_ManySmallFiles-10                                  123           9690994 ns/op
BenchmarkCopyFile_ManySmallFiles_InMemory-10                     4774060               253.8 ns/op
PASS
ok      github.com/nlepage/go-tarfs/benchmarks  6.694s
```

Benchmarks after:
```
$ go test -bench=.                
goos: darwin
goarch: arm64
pkg: github.com/nlepage/go-tarfs/benchmarks
BenchmarkOpenTarThenCopyFile_ManySmallFiles-10                        25          44993760 ns/op
BenchmarkOpenTarThenCopyFile_ManySmallFiles_InMemory-10               21          51123639 ns/op
BenchmarkCopyFile_ManySmallFiles-10                              1411468               855.9 ns/op
BenchmarkCopyFile_ManySmallFiles_InMemory-10                     4746981               249.9 ns/op
PASS
ok      github.com/nlepage/go-tarfs/benchmarks  6.612s
```