@joelanford 2023-07-22
Added another commit, which simplifies code further and seems to make some small gains in the benchmark:

```
$ go test -bench=.
goos: darwin
goarch: arm64
pkg: github.com/nlepage/go-tarfs/benchmarks
BenchmarkOpenTarThenCopyFile_ManySmallFiles-10                        25          44698328 ns/op
BenchmarkOpenTarThenCopyFile_ManySmallFiles_InMemory-10               28          40716345 ns/op
BenchmarkCopyFile_ManySmallFiles-10                              1400550               851.6 ns/op
BenchmarkCopyFile_ManySmallFiles_InMemory-10                     5445105               223.0 ns/op
PASS
ok      github.com/nlepage/go-tarfs/benchmarks  7.500s
```