@joelanford 2023-08-04
Updated before and after benchmarks:

Before:
```
$ go test ./benchmarks -bench=. -benchmem -memprofile memprofile.out
goos: darwin
goarch: arm64
pkg: github.com/nlepage/go-tarfs/benchmarks
BenchmarkOpenTarThenReadFile_ManySmallFiles-10                22          48194477 ns/op         8479395 B/op     163485 allocs/op
BenchmarkOpenTarThenReadFile_FewLargeFiles-10                261           4552260 ns/op        45079848 B/op        185 allocs/op
BenchmarkReadFile_ManySmallFiles-10                       259962              4424 ns/op            9338 B/op         18 allocs/op
BenchmarkReadFile_FewLargeFiles-10                           265           4455064 ns/op        45071809 B/op         21 allocs/op
PASS
ok      github.com/nlepage/go-tarfs/benchmarks  7.300s
```

After:
```
$ go test ./benchmarks -bench=. -benchmem -memprofile memprofile.out
goos: darwin
goarch: arm64
pkg: github.com/nlepage/go-tarfs/benchmarks
BenchmarkOpenTarThenReadFile_ManySmallFiles-10                        24          48131877 ns/op         8475510 B/op     163484 allocs/op
BenchmarkOpenTarThenReadFile_ManySmallFiles_DisableSeek-10            24          48353562 ns/op         8480806 B/op     163488 allocs/op
BenchmarkOpenTarThenReadFile_FewLargeFiles-10                       1352            763712 ns/op        45081608 B/op        186 allocs/op
BenchmarkOpenTarThenReadFile_FewLargeFiles_DisableSeek-10           1551            790632 ns/op        45081767 B/op        186 allocs/op
BenchmarkReadFile_ManySmallFiles-10                               356251              3331 ns/op            9291 B/op         17 allocs/op
BenchmarkReadFile_ManySmallFiles_DisableSeek-10                   346794              3379 ns/op            9291 B/op         17 allocs/op
BenchmarkReadFile_FewLargeFiles-10                                  1446            746129 ns/op        45071030 B/op         20 allocs/op
BenchmarkReadFile_FewLargeFiles_DisableSeek-10                      1700            753582 ns/op        45070915 B/op         20 allocs/op
BenchmarkOpenAndReadFile_ManySmallFiles-10                        204973              5760 ns/op           17668 B/op         21 allocs/op
BenchmarkOpenAndReadFile_ManySmallFiles_DisableSeek-10            269625              4218 ns/op            9339 B/op         18 allocs/op
BenchmarkOpenAndReadFile_FewLargeFiles-10                            166           7041059 ns/op        90136624 B/op         23 allocs/op
BenchmarkOpenAndReadFile_FewLargeFiles_DisableSeek-10                268           4464599 ns/op        45071708 B/op         21 allocs/op
PASS
ok      github.com/nlepage/go-tarfs/benchmarks  18.063s
```

Note the new `BenchmarkOpenAndReadFile` benchmarks which open the file and read it rather than call `ReadFile`. You can see that the `DisableSeek` variant uses half the memory and is generally faster to boot!